pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toppers"
version = "0.1.0"
description = "AI-powered Top 10 list video generator for YouTube Shorts"
authors = [
    {name = "Toppers AI", email = "noreply@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.203.1,<1.0.0",
    "google-cloud-storage>=2.10.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
toppers = "toppers.main:main"
